+++ tests/e2e/modules/helloworld_spec.js
@@ -6,32 +6,28 @@ describe("Test helloworld module", () => {
 	});
 
 	describe("helloworld set config text", () => {
-		beforeAll((done) => {
+		beforeAll(async () => {
 			helpers.startApplication("tests/configs/modules/helloworld/helloworld.js");
-			helpers.getDocument(done);
+			await helpers.getDocument();
 		});
 
-		it("Test message helloworld module", (done) => {
-			helpers.waitForElement(".helloworld").then((elem) => {
-				done();
-				expect(elem).not.toBe(null);
-				expect(elem.textContent).toContain("Test HelloWorld Module");
-			});
+		it("Test message helloworld module", async () => {
+			const elem = await helpers.waitForElement(".helloworld");
+			expect(elem).not.toBe(null);
+			expect(elem.textContent).toContain("Test HelloWorld Module");
 		});
 	});
 
 	describe("helloworld default config text", () => {
-		beforeAll((done) => {
+		beforeAll(async () => {
 			helpers.startApplication("tests/configs/modules/helloworld/helloworld_default.js");
-			helpers.getDocument(done);
+			await helpers.getDocument();
 		});
 
-		it("Test message helloworld module", (done) => {
-			helpers.waitForElement(".helloworld").then((elem) => {
-				done();
-				expect(elem).not.toBe(null);
-				expect(elem.textContent).toContain("Hello World!");
-			});
+		it("Test message helloworld module", async () => {
+			const elem = await helpers.waitForElement(".helloworld");
+			expect(elem).not.toBe(null);
+			expect(elem.textContent).toContain("Hello World!");
 		});
 	});
 });