+++ tests/e2e/modules/mocks/weather_current.js
@@ -4,7 +4,7 @@ const _ = require("lodash");
  * @param {object} extendedData extra data to add to the default mock data
  * @returns {string} mocked current weather data
  */
-function generateWeather(extendedData = {}) {
+const generateWeather = (extendedData = {}) => {
 	return JSON.stringify(
 		_.merge(
 			{},
@@ -59,6 +59,6 @@ function generateWeather(extendedData = {}) {
 			extendedData
 		)
 	);
-}
+};
 
 module.exports = generateWeather;