+++ tests/unit/classes/deprecated_spec.js
@@ -1,11 +1,11 @@
 const deprecated = require("../../../js/deprecated");
 
-describe("Deprecated", function () {
-	it("should be an object", function () {
+describe("Deprecated", () => {
+	it("should be an object", () => {
 		expect(typeof deprecated).toBe("object");
 	});
 
-	it("should contain configs array with deprecated options as strings", function () {
+	it("should contain configs array with deprecated options as strings", () => {
 		expect(Array.isArray(["deprecated.configs"])).toBe(true);
 		for (let option of deprecated.configs) {
 			expect(typeof option).toBe("string");